util/syspolicy: finish adding ts_omit_syspolicy build tags, tests
Fixes #16998
Updates #12614
